Use idiomatic comparisons with None.
It's not necessary to say "if x is not None:" or "if x is None:" - "if x:" and "if not x:" are just equivalent but more idiomatic.
In startflash, return early if the process is already running - allows the scope of the rest of the function to be taken outside the conditinal.