Use a combobox to select firmware type.
The corresponding addresses are stored in the 'userdata'. It's less direct, but allows the user to see a much friendlier option. It means that if there are only two correct addresses, the possibility of a user inserting incorrect values is removed (and the need for validation is similarly removed).

Related to this: reading the config file is now pulled out into a separate function that returns default values on any file read errors.